repackage evaluation into infrastructure
