+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/port/LoanEvaluationStatus.java
