--- src/loan-service/src/main/java/com/asc/loanservice/domain/loan/evaluation/LoanEvaluationAggregate.java
@@ -1,5 +1,0 @@
-package com.asc.loanservice.domain.loan.evaluation;
-
-public class LoanEvaluationAggregate {
-
-}
