+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/rule/CustomerAgeRule.java
@@ -1,11 +1,11 @@
 package com.asc.loanservice.infrastructure.rule;
 
 import com.asc.loanservice.contracts.LoanApplicationRequest;
-import com.asc.loanservice.domain.loan.evaluation.port.RulePort;
+import com.asc.loanservice.domain.loan.application.port.RulePort;
 import org.springframework.stereotype.Service;
 
 @Service
-public class MaximumInstallmentAmountRuleAdapter implements RulePort {
+public class CustomerAgeRuleAdapter implements RulePort {
 
     @Override
     public Boolean isValid(LoanApplicationRequest loanApplicationRequest) {
