+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/rule/Rule.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.domain.loan.evaluation.port;
+package com.asc.loanservice.domain.loan.application.port;
 
 import com.asc.loanservice.contracts.LoanApplicationRequest;
