+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/rule/port/DebtPort.java
@@ -1,7 +1,8 @@
-package com.asc.loanservice.domain.loan.evaluation.port;
+package com.asc.loanservice.domain.loan.application.port;
 
 import com.asc.loanservice.external.DebtCustomerResult;
 
 public interface DebtPort {
+
     DebtCustomerResult checkDebtClient(String customerTaxId);
 }
